# doc/api.yaml
openapi: 3.0.3
info:
  title: WASATEXT
  description: |-
    A webapp for communication.
  version: "1.0.0"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []

#login
paths:
  /session:
    post:
      tags: ["login"]
      security: []
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and then the identifier is returned.
        If the user exists the, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: Schema for entering username
              properties:
                name:
                  type: string
                  example: Trump
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                  description: The name of the user logging in or registering.
        required: true
      responses:
        '201':
          description: User login action successful
          content:
            application/json:
              schema:
                type: object
                description: response schema for sucessfull login
                properties:
                  identifier:
                    type: string
                    example: "abcdef012345"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                    description: unique id allocated to each user
                additionalProperties: false

#setMyUserName
...
  /conversations:
    get:
      tags: ["conversations"]
      summary: Get user's conversations
      description: |-
        Retrieves a list of conversations the user is part of,
        sorted by the last message time.
      operationId: getMyConversations
      responses:
        '200':
          description: List of Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  description: array to conversations user is part of
                  properties:
                    conversationId:
                      type: string
                      example: "12345"
                      pattern: '^.*?$'
                      minLength: 3
                      maxLength: 16
                      description: unique identifier to each conversation
                    name:
                      type: string
                      example: "DonaldTrump or Groupname"
                      pattern: '^.*?$'
                      minLength: 3
                      maxLength: 16
                      description: Name of person or group
                    photoUrl:
                      type: string
                      example: "https://example.com/photo.jpg"
                      pattern: '^.*?$'
                      minLength: 3
                      maxLength: 1000
                      description: Url of group or person photo
                    lastmessage:
                      type: object
                      description: the last message receved by user
                      properties:
                        text:
                          type: string
                          example: "Hello Im under the Water"
                          pattern: '^.*?$'
                          minLength: 1
                          maxLength: 1000
                          description: content of text message
                        type:
                          type: string
                          enum: [text,photo]
                          example: "text"
                          pattern: '^.*?$'
                          minLength: 4
                          maxLength: 16
                          description: type of last message

                        timestamp:
                          type: string
                          format: date-time
                          example: "2012-12-12T12:00:00Z" 
                          pattern: '^.*?$'
                          minLength: 3
                          maxLength: 40
                          description: Timestamp of last message
                        sender:
                          type: string
                          example: "Harris"
                          pattern: '^.*?$'
                          minLength: 3
                          maxLength: 16
                          description: Username of the person who sent the message
                  additionalProperties: false
                minItems: 0
                maxItems: 100  
    
        '401':
          description: Authentication Required
          content:
            application/json:
              schema:
                type: object
                description: response scehma for Authentication required
                properties:
                  error:
                    type: string
                    example: "Authentication required."
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 60
                    description: Auth required error message
                additionalProperties: false

#getConversation
  /conversations/{id}:
    get:
      tags: ["conversation"]
      summary: Gets Conversation Details
      description: Retrieves the messages and metadata from a specific conversation
      operationId: getConversation
      parameters:
        - in: path
          name: id 
          required: true
          schema:
            type: string
            pattern: '^.*?$'
            minLength: 3
            maxLength: 16
            description: Unique identifier of the Conversation
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 50
            description: Maximum number of messages to retireve
        - in: query
          name: offset
          required: false
          schema: 
            type: integer
            default: 0
            description: Number of Messages to Skip(for pagination)
      responses:
        '200':
          description: conversation details
          content: 
            application/json:
              schema:
                type: object
                description: basic conversation details of user
                properties: 
                  conversationId:
                    type: string
                    example: "6969"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                    description: Unique identifier of the conversation
                  name:
                    type: string
                    example: "Groupname or username"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                    description: Name of the Group or User
                  photoUrl:
                    type: string
                    example: "https://example.com/photo.jpg"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 1000
                    description: url of the photo
                  messages:
                    type: array
                    items:
                      type: object
                      description: array of messages receved by user
                      properties: 
                        messageId:
                          type: string
                          example: "23233"
                          pattern: '^.*?$'
                          minLength: 3
                          maxLength: 16
                          description: unique identifier of the message
                        text:
                          type: string
                          example: "Halo welt!!!"
                          pattern: '^.*?$'
                          minLength: 1
                          maxLength: 1000
                          description: Content of the message
                        type:
                          type: string
                          enum: [text,photo]
                          example: "text"
                          pattern: '^.*?$'
                          minLength: 4
                          maxLength: 16
                          description: type of content in the message
                        timestamp:
                          type: string
                          format: date-time
                          example: "2012-12-12T12:00:00Z"
                          pattern: '^.*?$'
                          minLength: 3
                          maxLength: 40
                          description: time stamp of the message
                        sender:
                          type: string
                          example: "Harris"
                          pattern: '^.*?$'
                          minLength: 3
                          maxLength: 16
                          description: usernme of the sender
                    minItems: 1
                    maxItems: 100
                    description: Contents and metadata of message
        '401':
          description: Authentication required
          content:
            application/json:
              schema: 
                type: object
                description: response schema for authentication required
                properties:
                  error:
                    type: string
                    example: "Authentication Required"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 60
                    description: Auth required Error msg
        '404':
          description: Conversation not Found
          content:
            application/json:
              schema:
                type: object
                description: response schema for conversation not found
                properties:
                  error:
                    type: string
                    example: "Conversation not found."
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 60
                    description: conversation not found message
              
#sendMessage
...
  /groups:
    post:
      summary: Create a new group
      description: Allows a user to create a new group with any number of other WASAText users.
      operationId: createGroup
      tags: ["group"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: schema to create a group
              properties:
                groupName:
                  type: string
                  example: "Friends Group"
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                  description: name of the group
                members:
                  type: array
                  items:
                    type: string
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                    description: Username of the user
                  example: ["user1", "user2", "user3"]
                  pattern: '^.*?$'
                  minItems: 1
                  maxItems: 100
                  description: usernames of the members of the group
              required:
                - groupName
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                type: object
                description: response to group created sucessfully
                properties:
                  groupId:
                    type: string
                    example: "group123"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                    description: unique group identifier
                  groupName:
                    type: string
                    example: "Friends Group"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                    description: Name of the Group
                  admin:
                    type: string
                    example: "user1"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                    description: username of the Admin of the group
                  members:
                    type: array
                    items:
                      type: string
                    example: ["user1", "user2", "user3"]
                    minItems: 1
                    maxItems: 100
                    description: usernames of the members of the group
                  photoUrl:
                    type: string
                    example: "https://example.com/group-photo.jpg"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 1000
                    description: Url of the photo
        '401':
          description: Authentication required
        '400':
          description: Invalid request data

#setGroupName
  /groups/{groupId}/name:
    put:
      summary: Update group name
      description: Allows the admin of a group to update its name.
      operationId: setGroupName
      tags: ["group"]
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
            pattern: '^.*?$'
            minLength: 3
            maxLength: 16
          description: ID of the group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: schema to change group name
              properties:
                newName:
                  type: string
                  example: "New Group Name"
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                  description: New name of the Group
              required:
                - newName
      responses:
        '200':
          description: Group name updated successfully
          content:
            application/json:
              schema:
                type: object
                description: response schema for sucessfully changing grouo name
                properties:
                  message:
                    type: string
                    example: "Group name updated successfully."
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 60
                    description: 200 success message
                  groupName:
                    type: string
                    example: "Trumpgroup2"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                    description: new group name
        '401':
          description: Authentication required
        '403':
          description: Only the admin can update the group name
        '404':
          description: Group not found

#setGroupPhoto
  /groups/{groupId}/photo:
    put:
      summary: Update group photo
      description: Allows the admin of a group to update its photo.
      operationId: setGroupPhoto
      tags: ["group"]
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
            pattern: '^.*?$'
            minLength: 3
            maxLength: 16
          description: ID of the group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: photo of group
              properties:
                photoUrl:
                  type: string
                  example: "https://example.com/new-group-photo.jpg"
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 1000
                  description: Url of the Photo
              required:
                - photoUrl
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: response to sucess in group photo change
                properties:
                  message:
                    type: string
                    example: "Group photo updated successfully."
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 60
                    description: 200 sucess message
                  photoUrl:
                    type: string
                    example: "https://example.com/new-group-photo.jpg"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 1000
                    description: Url of the photo
        '401':
          description: Authentication required
        '403':
          description: Only the admin can update the group photo
        '404':
          description: Group not found

#add people to group(addToGroup)
...